requirements.txt
Flask==3.0.0
google-genai==0.2.2
python-dotenv==1.0.0
SQLAlchemy==2.0.37
alembic==1.13.2
pytest==8.3.4
rq==1.16.2
redis==5.0.7
